runpod_templates/requirements_image.txt
runpod==1.7.7
torch>=2.2.0
diffusers>=0.31.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.3
Pillow>=10.4.0
boto3>=1.34.0
sentencepiece>=0.2.0
